Optimize icon (#13)
* feat: Add transpilePackages option to next.config

This commit adds the `transpilePackages` option to the `next.config` file. The `transpilePackages` array includes the package `lucide-react` to be transpiled during the build process. This change is necessary to ensure that the `lucide-react` package is properly transpiled and included in the final build.

* fix

This commit updates the logo component by replacing the Lucide icon with the Triangle icon from the Lucide React package. It also updates the mode toggle component by replacing the Sun and Moon icons with the corresponding icons from the Lucide React package. These changes improve the consistency and visual appeal of the components.

* fix

* fix

* fix